Add Solaris, HP-UX normal and trusted-mode password resets
The HP-UX operating system can run in what's known as "trusted mode,"
which includes, among other things, HP's own take on the "shadow" idea.
The /tcb/files/auth/u/username file contains a database of keys and
values specific to that user, among which are the password hash with
a key of u_pwd, and the last successful password change time in
seconds since epoch with a key of u_succhg.

This commit adds support for HP-UX trusted mode, as well as
conventional /etc/shadow mode for HP-UX and Solaris which works in
exactly the same way as it does on Linux.