Deal with unusual but valid formats in TCB files
A well-patched HP-UX machine will have trailing colons at the end of
each TCB file line, just before the continuation backslash, but this
is not mandatory. Just a backslash, or a key ending the entry at
the end of the file, is also valid. Modified TCB editing to deal with
those corner cases.