Testing and validation of AIX promises
The sequence of promises in each section makes a difference. If
the password is set, then the set_user_password is asserted, but
we can't set the change time until _since_epoch is defined, which
won't occur until the second pass. So the time setting needs to
be moved ahead of the hash setting to insure that the _since_epoch
is available before it is evaluated.